+++ packages/orq-rc/src/orq_ai_sdk/models/filegetop.py
@@ -63,5 +63,5 @@ class FileGetResponseBody(BaseModel):
     workspace_id: str
     r"""The id of the resource"""
 
-    created: Optional[datetime] = parse_datetime("2025-09-25T06:48:42.376Z")
+    created: Optional[datetime] = parse_datetime("2025-09-25T12:10:32.801Z")
     r"""The date and time the resource was created"""