+++ packages/orq-rc/src/orq_ai_sdk/models/fileuploadop.py
@@ -101,5 +101,5 @@ class FileUploadResponseBody(BaseModel):
     workspace_id: str
     r"""The id of the resource"""
 
-    created: Optional[datetime] = parse_datetime("2025-09-25T06:48:42.376Z")
+    created: Optional[datetime] = parse_datetime("2025-09-25T12:10:32.801Z")
     r"""The date and time the resource was created"""