+++ packages/orq-rc/src/orq_ai_sdk/models/getagentop.py
@@ -247,7 +247,7 @@ class GetAgentKnowledgeBases(BaseModel):
     configuration: GetAgentConfiguration
     r"""Defines the configuration settings which can either be for a user message or a text entry."""
 
-    id: Optional[str] = "01K5ZT3CQCJNRJAK5JM21YBYM3"
+    id: Optional[str] = "01K60CGQ2DXANBB2SN1A3NQ2CT"
     r"""The id of the resource"""