+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -132,7 +132,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K5ZT3CV84M7WC2SHW3YJHN1W"
+        "01K60CGQ6KMNCZ11M6ZTSX707R"
     )
     r"""The id of the resource"""