+++ packages/orq-rc/RELEASES.md
@@ -2899,3 +2899,13 @@ Based on:
 - [python v3.13.0-rc.36] packages/orq-rc
 ### Releases
 - [PyPI v3.13.0-rc.36] https://pypi.org/project/orq-ai-sdk/3.13.0-rc.36 - packages/orq-rc
+
+## 2025-09-25 12:12:07
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.628.4 (2.716.5) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.13.0-rc.37] packages/orq-rc
+### Releases
+- [PyPI v3.13.0-rc.37] https://pypi.org/project/orq-ai-sdk/3.13.0-rc.37 - packages/orq-rc